requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Flask-CORS==4.0.0
PyJWT==2.8.0
pytest==7.4.0
pytest-cov==4.1.0
